# examples/basic/docs/openapi.json
{
  "info": {
    "description": "RESTful API for managing users",
    "summary": "Basic REST API",
    "title": "User API",
    "version": "1.1"
  },
  "openapi": "3.0.0",
  "paths": {
    "/api/v1/users": {
      "get": {
        "description": "Retrieve a list of resources",
        "responses": {
          "204": {
            "description": "No Content"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "properties": {
                    "code": {
                      "example": "bad_request",
                      "type": "string"
                    },
                    "message": {
                      "example": "Failed to parse request",
                      "type": "string"
                    },
                    "status": {
                      "example": 400,
                      "type": "integer"
                    }
                  },
                  "required": [
                    "status",
                    "code",
                    "message"
                  ],
                  "type": "object"
                }
              }
            },
            "description": "Bad Request"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "properties": {
                    "code": {
                      "example": "validation",
                      "type": "string"
                    },
                    "errors": {
                      "items": {
                        "properties": {
                          "field": {
                            "type": "string"
                          },
                          "message": {
                            "type": "string"
                          }
                        },
                        "type": "object"
                      },
                      "type": "array"
                    },
                    "message": {
                      "example": "Validation failed",
                      "type": "string"
                    },
                    "status": {
                      "example": 422,
                      "type": "integer"
                    }
                  },
                  "required": [
                    "status",
                    "code",
                    "message"
                  ],
                  "type": "object"
                }
              }
            },
            "description": "Validation Error"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "properties": {
                    "code": {
                      "example": "internal",
                      "type": "string"
                    },
                    "message": {
                      "example": "Internal server error",
                      "type": "string"
                    },
                    "status": {
                      "example": 500,
                      "type": "integer"
                    }
                  },
                  "required": [
                    "status",
                    "code",
                    "message"
                  ],
                  "type": "object"
                }
              }
            },
            "description": "Internal Server Error"
          }
        },
        "summary": "List resources",
        "tags": [
          "User Management v1"
        ],
        "title": "List"
      },
      "post": {
        "description": "Create a new resource",
        "responses": {
          "204": {
            "description": "No Content"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "properties": {
                    "code": {
                      "example": "bad_request",
                      "type": "string"
                    },
                    "message": {
                      "example": "Failed to parse request",
                      "type": "string"
                    },
                    "status": {
                      "example": 400,
                      "type": "integer"
                    }
                  },
                  "required": [
                    "status",
                    "code",
                    "message"
                  ],
                  "type": "object"
                }
              }
            },
            "description": "Bad Request"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "properties": {
                    "code": {
                      "example": "validation",
                      "type": "string"
                    },
                    "errors": {
                      "items": {
                        "properties": {
                          "field": {
                            "type": "string"
                          },
                          "message": {
                            "type": "string"
                          }
                        },
                        "type": "object"
                      },
                      "type": "array"
                    },
                    "message": {
                      "example": "Validation failed",
                      "type": "string"
                    },
                    "status": {
                      "example": 422,
                      "type": "integer"
                    }
                  },
                  "required": [
                    "status",
                    "code",
                    "message"
                  ],
                  "type": "object"
                }
              }
            },
            "description": "Validation Error"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "properties": {
                    "code": {
                      "example": "internal",
                      "type": "string"
                    },
                    "message": {
                      "example": "Internal server error",
                      "type": "string"
                    },
                    "status": {
                      "example": 500,
                      "type": "integer"
                    }
                  },
                  "required": [
                    "status",
                    "code",
                    "message"
                  ],
                  "type": "object"
                }
              }
            },
            "description": "Internal Server Error"
          }
        },
        "summary": "Create a new resource",
        "tags": [
          "User Management v1"
        ],
        "title": "Create"
      }
    },
    "/api/v1/users/{id}": {
      "delete": {
        "description": "Delete a resource by ID",
        "parameters": [
          {
            "description": "id parameter",
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "No Content"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "properties": {
                    "code": {
                      "example": "bad_request",
                      "type": "string"
                    },
                    "message": {
                      "example": "Failed to parse request",
                      "type": "string"
                    },
                    "status": {
                      "example": 400,
                      "type": "integer"
                    }
                  },
                  "required": [
                    "status",
                    "code",
                    "message"
                  ],
                  "type": "object"
                }
              }
            },
            "description": "Bad Request"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "properties": {
                    "code": {
                      "example": "validation",
                      "type": "string"
                    },
                    "errors": {
                      "items": {
                        "properties": {
                          "field": {
                            "type": "string"
                          },
                          "message": {
                            "type": "string"
                          }
                        },
                        "type": "object"
                      },
                      "type": "array"
                    },
                    "message": {
                      "example": "Validation failed",
                      "type": "string"
                    },
                    "status": {
                      "example": 422,
                      "type": "integer"
                    }
                  },
                  "required": [
                    "status",
                    "code",
                    "message"
                  ],
                  "type": "object"
                }
              }
            },
            "description": "Validation Error"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "properties": {
                    "code": {
                      "example": "internal",
                      "type": "string"
                    },
                    "message": {
                      "example": "Internal server error",
                      "type": "string"
                    },
                    "status": {
                      "example": 500,
                      "type": "integer"
                    }
                  },
                  "required": [
                    "status",
                    "code",
                    "message"
                  ],
                  "type": "object"
                }
              }
            },
            "description": "Internal Server Error"
          }
        },
        "summary": "Delete a resource by ID",
        "tags": [
          "User Management v1"
        ],
        "title": "Delete"
      },
      "get": {
        "description": "Look up a resource",
        "parameters": [
          {
            "description": "",
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "No Content"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "properties": {
                    "code": {
                      "example": "bad_request",
                      "type": "string"
                    },
                    "message": {
                      "example": "Failed to parse request",
                      "type": "string"
                    },
                    "status": {
                      "example": 400,
                      "type": "integer"
                    }
                  },
                  "required": [
                    "status",
                    "code",
                    "message"
                  ],
                  "type": "object"
                }
              }
            },
            "description": "Bad Request"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "properties": {
                    "code": {
                      "example": "validation",
                      "type": "string"
                    },
                    "errors": {
                      "items": {
                        "properties": {
                          "field": {
                            "type": "string"
                          },
                          "message": {
                            "type": "string"
                          }
                        },
                        "type": "object"
                      },
                      "type": "array"
                    },
                    "message": {
                      "example": "Validation failed",
                      "type": "string"
                    },
                    "status": {
                      "example": 422,
                      "type": "integer"
                    }
                  },
                  "required": [
                    "status",
                    "code",
                    "message"
                  ],
                  "type": "object"
                }
              }
            },
            "description": "Validation Error"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "properties": {
                    "code": {
                      "example": "internal",
                      "type": "string"
                    },
                    "message": {
                      "example": "Internal server error",
                      "type": "string"
                    },
                    "status": {
                      "example": 500,
                      "type": "integer"
                    }
                  },
                  "required": [
                    "status",
                    "code",
                    "message"
                  ],
                  "type": "object"
                }
              }
            },
            "description": "Internal Server Error"
          }
        },
        "summary": "Get a resource",
        "tags": [
          "User Management v1"
        ],
        "title": "Get"
      },
      "patch": {
        "description": "Update a resource by ID",
        "parameters": [
          {
            "description": "",
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "No Content"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "properties": {
                    "code": {
                      "example": "bad_request",
                      "type": "string"
                    },
                    "message": {
                      "example": "Failed to parse request",
                      "type": "string"
                    },
                    "status": {
                      "example": 400,
                      "type": "integer"
                    }
                  },
                  "required": [
                    "status",
                    "code",
                    "message"
                  ],
                  "type": "object"
                }
              }
            },
            "description": "Bad Request"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "properties": {
                    "code": {
                      "example": "validation",
                      "type": "string"
                    },
                    "errors": {
                      "items": {
                        "properties": {
                          "field": {
                            "type": "string"
                          },
                          "message": {
                            "type": "string"
                          }
                        },
                        "type": "object"
                      },
                      "type": "array"
                    },
                    "message": {
                      "example": "Validation failed",
                      "type": "string"
                    },
                    "status": {
                      "example": 422,
                      "type": "integer"
                    }
                  },
                  "required": [
                    "status",
                    "code",
                    "message"
                  ],
                  "type": "object"
                }
              }
            },
            "description": "Validation Error"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "properties": {
                    "code": {
                      "example": "internal",
                      "type": "string"
                    },
                    "message": {
                      "example": "Internal server error",
                      "type": "string"
                    },
                    "status": {
                      "example": 500,
                      "type": "integer"
                    }
                  },
                  "required": [
                    "status",
                    "code",
                    "message"
                  ],
                  "type": "object"
                }
              }
            },
            "description": "Internal Server Error"
          }
        },
        "summary": "Update a resource by ID",
        "tags": [
          "User Management v1"
        ],
        "title": "Update"
      }
    },
...
  },
  "servers": [
    {
      "description": "Dev Server",
      "url": "http://localhost:8080",
      "variables": null
    }
  ],
...
}